requirements.txt
python-dotenv
psycopg2-binary==2.9.9
Flask==2.0.2
SQLAlchemy==2.0.34
boto3==1.35.34
python-dotenv==1.0.1
botocore==1.35.34
colorlog==6.4.1
awscli
psutil